Customize color pallete and change demo button color to secondary and rename boats to boat
