+++ app/models/__init__.py
@@ -1,5 +1,5 @@
 from .db import db
 from .user import User
-from .boats import Boat
+from .boat import Boat
 from .reservation import Reservation
 from .media import Media
